+++ src/main/java/com/phantomwing/eastersdelight/datagen/ModPoiTypeTagsProvider.java
@@ -1,0 +1,25 @@
+package com.phantomwing.eastersdelight.datagen;
+
+import com.phantomwing.eastersdelight.EastersDelight;
+import com.phantomwing.eastersdelight.villager.ModVillagers;
+import net.minecraft.core.HolderLookup;
+import net.minecraft.data.PackOutput;
+import net.minecraft.data.tags.PoiTypeTagsProvider;
+import net.minecraft.tags.PoiTypeTags;
+import net.neoforged.neoforge.common.data.ExistingFileHelper;
+import org.jetbrains.annotations.Nullable;
+
+import java.util.Objects;
+import java.util.concurrent.CompletableFuture;
+
+public class ModPoiTypeTagsProvider extends PoiTypeTagsProvider {
+    public ModPoiTypeTagsProvider(PackOutput output, CompletableFuture<HolderLookup.Provider> lookupProvider, @Nullable ExistingFileHelper existingFileHelper) {
+        super(output, lookupProvider, EastersDelight.MOD_ID, existingFileHelper);
+    }
+
+    @Override
+    protected void addTags(HolderLookup.Provider provider) {
+        this.tag(PoiTypeTags.ACQUIRABLE_JOB_SITE)
+                .addOptional(Objects.requireNonNull(ModVillagers.EGG_BUNNY_POI.getKey()).location());
+    }
+}