+++ src/main/java/com/phantomwing/eastersdelight/item/custom/DyedEggItem.java
@@ -1,9 +1,37 @@
 package com.phantomwing.eastersdelight.item.custom;
 
-import net.minecraft.world.item.Item;
+import com.phantomwing.eastersdelight.EastersDelight;
+import com.phantomwing.eastersdelight.component.EggPattern;
+import com.phantomwing.eastersdelight.component.ModDataComponents;
+import net.minecraft.network.chat.Component;
+import net.minecraft.network.chat.MutableComponent;
+import net.minecraft.world.item.BlockItem;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.TooltipFlag;
+import net.minecraft.world.level.block.Block;
+import org.jetbrains.annotations.NotNull;
 
-public class DyedEggItem extends Item {
-    public DyedEggItem(Properties properties) {
-        super(properties);
+import java.util.List;
+
+public class DyedEggItem extends BlockItem {
+    public DyedEggItem(Block block, Properties properties) {
+        super(block, properties);
+    }
+
+    @Override
+    public void appendHoverText(ItemStack stack, @NotNull TooltipContext context, @NotNull List<Component> tooltipComponents, @NotNull TooltipFlag tooltipFlag) {
+        if (stack.has(ModDataComponents.EGG_PATTERN)) {
+            EggPattern pattern = stack.get(ModDataComponents.EGG_PATTERN);
+
+            if (pattern == null) {
+                return;
+            }
+
+            MutableComponent patternTooltip = Component.translatable(EastersDelight.MOD_ID + ".tooltip.egg_pattern." + pattern.getName());
+
+            tooltipComponents.add(patternTooltip);
+        }
+
+        super.appendHoverText(stack, context, tooltipComponents, tooltipFlag);
     }
 }