+++ src/main/java/com/phantomwing/eastersdelight/tags/CommonTags.java
@@ -25,6 +25,7 @@ public class CommonTags {
     public static final TagKey<Item> FOODS_BOILED_EGG = commonItemTag("foods/boiled_egg");
 
     public static final TagKey<Item> TOOLS_KNIFE = commonItemTag("tools/knife");
+    public static final TagKey<Block> MINEABLE_WITH_KNIFE = commonBlockTag("mineable/knife");
 
     private static TagKey<Block> commonBlockTag(String path) {
         return BlockTags.create(ResourceLocation.fromNamespaceAndPath("c", path));