+++ changelog.md
@@ -1,3 +1,13 @@
+# 1.1.0
+### Additions
+- Dyed eggs are now placeable!
+- Add crafting recipe for 'simple' dyed eggs (without patterns)
+  - Just combine a boiled egg with any dye
+- Added Ukrainian translations (thank you Sun_Phoen1x!)
+
+### Fixes
+- Fixed Baked Cod Stew recipe override not having a Bowl container
+
 # 1.0.1
 ### Fixes
 - `Bunny Cookie` is now a poisonous food item for Parrots (as intended)